Created a separate class for the Obstacles
I moved the struct Obstacle as a class to a separate .cs file with all
its methods as well.
I added an additional folder Scores to store the TopScore.txt file.
Currently we are getting an unhandled exception if we try to parse from
an empty file. Writing a 0 at the first line of the file eliminates the
problem.