Added - Reading initial configuration values from file
GamePlay.WindowsSize() method was changed to accept a single argument -
string(path to the config file). We then have an exception handling in
the method similar to the one we are using with the TopScore file.
The playFieldWidth variable was moved as a static public variable in the
GamePlay class.
Some changes in the FlappyGriffin.cs - we initialize the Window before
the Griffin - in order to have the playFieldWidth variable initialized
by the WindowSize() method.